chore: update docs
