+++ src/app/Components/PriceRange/constants.ts
@@ -1,3 +1,17 @@
+import { SelectOption } from "app/Components/Select"
+
 export const DEFAULT_RANGE = [0, 50000]
 export const RANGE_DOT_SIZE = 32
 export const DEFAULT_PRICE_RANGE = "*-*"
+
+export const PRICE_BUCKETS: Array<SelectOption<string>> = [
+  { label: "Select a price range", value: "" },
+  { label: "Under $500", value: "-1:500" },
+  { label: "Under $2,500", value: "-1:2500" },
+  { label: "Under $5,000", value: "-1:5000" },
+  { label: "Under $10,000", value: "-1:10000" },
+  { label: "Under $25,000", value: "-1:25000" },
+  { label: "Under $50,000", value: "-1:50000" },
+  { label: "Under $100,000", value: "-1:100000" },
+  { label: "No budget in mind", value: "-1:1000000000000" },
+]
